chore: rename to account kit boilerplate
