+++ android/app/src/main/java/com/accountkitboilerplate/MainActivity.java
@@ -1,4 +1,4 @@
-package com.rntypescriptboilerplate;
+package com.AccountKitBoilerplate;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -23,7 +23,7 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "RNTypescriptBoilerplate";
+    return "AccountKitBoilerplate";
   }
 
   /**
