+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "RNTypescriptBoilerplate",
-  "displayName": "RNTypescriptBoilerplate"
+  "name": "AccountKitBoilerplate",
+  "displayName": "AccountKitBoilerplate"
 }
