+++ ios/AccountKitBoilerplate.xcodeproj/xcshareddata/xcschemes/AccountKitBoilerplate.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "RNTypescriptBoilerplate.app"
-               BlueprintName = "RNTypescriptBoilerplate"
-               ReferencedContainer = "container:RNTypescriptBoilerplate.xcodeproj">
+               BuildableName = "AccountKitBoilerplate.app"
+               BlueprintName = "AccountKitBoilerplate"
+               ReferencedContainer = "container:AccountKitBoilerplate.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "RNTypescriptBoilerplateTests.xctest"
-               BlueprintName = "RNTypescriptBoilerplateTests"
-               ReferencedContainer = "container:RNTypescriptBoilerplate.xcodeproj">
+               BuildableName = "AccountKitBoilerplateTests.xctest"
+               BlueprintName = "AccountKitBoilerplateTests"
+               ReferencedContainer = "container:AccountKitBoilerplate.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "RNTypescriptBoilerplate.app"
-            BlueprintName = "RNTypescriptBoilerplate"
-            ReferencedContainer = "container:RNTypescriptBoilerplate.xcodeproj">
+            BuildableName = "AccountKitBoilerplate.app"
+            BlueprintName = "AccountKitBoilerplate"
+            ReferencedContainer = "container:AccountKitBoilerplate.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "RNTypescriptBoilerplate.app"
-            BlueprintName = "RNTypescriptBoilerplate"
-            ReferencedContainer = "container:RNTypescriptBoilerplate.xcodeproj">
+            BuildableName = "AccountKitBoilerplate.app"
+            BlueprintName = "AccountKitBoilerplate"
+            ReferencedContainer = "container:AccountKitBoilerplate.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
