+++ ios/AccountKitBoilerplate.xcworkspace/contents.xcworkspacedata
@@ -2,9 +2,12 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:RNTypescriptBoilerplate.xcodeproj">
+      location = "group:AccountKitBoilerplate.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:AccountKitBoilerplate.xcodeproj">
    </FileRef>
 </Workspace>
