+++ ios/AccountKitBoilerplate.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
