+++ ios/AccountKitBoilerplate/AppDelegate.h
