+++ ios/AccountKitBoilerplate/AppDelegate.mm
@@ -8,7 +8,7 @@
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
 {
-  self.moduleName = @"RNTypescriptBoilerplate";
+  self.moduleName = @"AccountKitBoilerplate";
   // You can add your custom initial props in the dictionary below.
   // They will be passed down to the ViewController used by React Native.
   self.initialProps = @{};
