+++ ios/AccountKitBoilerplate/Images.xcassets/AppIcon.appiconset/Contents.json
