+++ ios/AccountKitBoilerplate/Images.xcassets/Contents.json
