+++ ios/AccountKitBoilerplate/Images.xcassets/LaunchScreen.imageset/Contents.json
