+++ ios/AccountKitBoilerplateTests/Info.plist
